--- src/pages/Post/PostDetail/components/MessagesPageStyle.js
@@ -1,19 +1,0 @@
-import { css } from "@emotion/react";
-
-export const ButtonGroupStyle = css`
-  display: flex;
-  gap: 10px;
-  margin: 16px 0 16px 0;
-`;
-
-export const ErrorTextStyle = css`
-  color: red;
-  margin-top: 12px;
-  font-size: 14px;
-  text-align: center;
-`;
-
-export const ObserverSpacerStyle = css`
-  height: 1px;
-  margin-top: 40px;
-`;
